Big patch. We no longer keep track of or "mirror" HTML attributes, CSS properties and DOM events ("visual state and data") on the server.
The reasons for storing visual state on the server no longer apply. I no longer share a single widget/View instance across multiple VIEWPORTs, but instead use MVC (one Model --many--> Views) to effectively do the same thing.

Though, a problem still remains in that the user might add a widget/View to the VIEWPORT, remove it, then add it back again; the visual state and data will then be lost if not kept track of on the server. I've solved this by "lazily" removing widget instances in the DOM based on GC+finalization on the server end. That is, the widgets are not "really" removed from the DOM, they are just hidden until we're sure they'll be gone for good.