We can't separate operations based on type ("DOM event") and name ("click") only.
See, e.g., the TEXT-INPUT widget for an example why. It creates a new event type based on the "already occupied" event name "blur".

To solve this we separate things in the DELAYED-OPERATIONS slot based on the actual name of the Lisp side API function as this is surely unique.